+++ category_encoders/hashing.py
@@ -9,6 +9,7 @@
 import copy
 from sklearn.base import BaseEstimator, TransformerMixin
 import pandas as pd
+from category_encoders.utils import get_obj_cols
 
 __author__ = 'willmcginnis'
 
@@ -100,6 +101,15 @@ class HashingEncoder(BaseEstimator, TransformerMixin):
         :return:
         """
 
+        # if the input dataset isn't already a dataframe, convert it to one (using default column names)
+        if not isinstance(X, pd.DataFrame):
+            X = pd.DataFrame(X)
+
+        # if columns aren't passed, just use every string column
+        if self.cols is None:
+            self.cols = get_obj_cols(X)
+
+        # drop all output columns with 0 variance.
         if self.drop_invariant:
             self.drop_cols = []
             X_temp = self.transform(X)