+++ category_encoders/polynomial.py
@@ -10,7 +10,7 @@ import copy
 import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 from patsy.highlevel import dmatrix
-from category_encoders.ordinal import OrdinalEncoder
+from category_encoders.utils import get_obj_cols
 
 __author__ = 'willmcginnis'
 
@@ -68,6 +68,15 @@ class PolynomialEncoder(BaseEstimator, TransformerMixin):
         :return:
         """
 
+        # if the input dataset isn't already a dataframe, convert it to one (using default column names)
+        if not isinstance(X, pd.DataFrame):
+            X = pd.DataFrame(X)
+
+        # if columns aren't passed, just use every string column
+        if self.cols is None:
+            self.cols = get_obj_cols(X)
+
+        # drop all output columns with 0 variance.
         if self.drop_invariant:
             self.drop_cols = []
             X_temp = self.transform(X)